update
created objects in vita and progress, created store arrays and methods, not working because these posts don't exist yet. Had some issues with understanding that mounted block, so I'd appreciate another explanation tomorrow. Also not sure if wee need an array for progress, need a bit of explanation on the purpose of that option